A better fix to updates.
If the memory information changes but there aren't any X events, the new
values were calculated but the image wasn't being updated. I added a call to
check for X events which, I discovered, solved my issue. All I really needed
was to perform an XFlush() after updating the image to push the new drawing
out to the window. My previous fix succeeded because there is an implicit
flush with the XNextEvent() call.